adding authentication

+++ backend/controllers/goalController.js
@@ -1,15 +1,14 @@
 const Goal = require("../models/goalModel")
 const asyncHandler = require("express-async-handler")
-
+const User = require("../models/userModel")
 
 //@desc     Get Goals
-//@route    GET /api/goals
+//@route    GET /api/goals 
 //@access   public
 const getGoals = asyncHandler(async (req, res, next) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
-}
-)
+})
 
 
 //@desc     Creat new goal
@@ -23,7 +22,8 @@ const setGoal = asyncHandler(async (req, res, next) => {
         throw new Error("Please add a text field")
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -40,7 +40,17 @@ const updateGoal = asyncHandler(async (req, res, next) => {
         res.status(400)
         throw new Error("Goal not found")
     }
-
+    const user = await User.findById(req.user.id)
+    //check for user
+    if (!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+    //Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true
     })
@@ -60,7 +70,18 @@ const deleteGoal = asyncHandler(async (req, res, next) => {
         res.status(400)
         throw new Error("Goal not found")
     }
+    const user = await User.findById(req.user.id)
+    //check for user
+    if (!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
 
+    //Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
     await Goal.findByIdAndDelete(req.params.id)
 
     res.status(200).json({
